update tropo trigger script
